+++ src/Storage/Storage.Management.Test/ScenarioTests/StorageFileTests.cs
@@ -50,6 +50,13 @@ namespace Microsoft.Azure.Commands.Management.Storage.Test.ScenarioTests
         public void TestStorageFileShareGetUsage()
         {
             TestController.NewInstance.RunPsTest(_logger, "Test-StorageFileShareGetUsage");
-        } 
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestFileServiceProperties()
+        {
+            TestController.NewInstance.RunPsTest(_logger, "Test-FileServiceProperties");
+        }
     }
 }