+++ src/Storage/Storage.Management/Models/PSDataPolicy.cs
@@ -212,6 +212,7 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
         public PSDateAfterModification TierToCool { get; set; }
         public PSDateAfterModification TierToArchive { get; set; }
         public PSDateAfterModification Delete { get; set; }
+        public bool? EnableAutoTierToHotFromCool { get; set; }
 
         public PSManagementPolicyBaseBlob()
         { }
@@ -221,6 +222,7 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
             this.TierToCool = blobAction.TierToCool is null ? null : new PSDateAfterModification(blobAction.TierToCool);
             this.TierToArchive = blobAction.TierToArchive is null ? null : new PSDateAfterModification(blobAction.TierToArchive);
             this.Delete = blobAction.Delete is null ? null : new PSDateAfterModification(blobAction.Delete);
+            this.EnableAutoTierToHotFromCool = blobAction.EnableAutoTierToHotFromCool;
         }
         public ManagementPolicyBaseBlob ParseManagementPolicyBaseBlob()
         {
@@ -228,7 +230,8 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
             {
                 TierToCool = this.TierToCool is null ? null : this.TierToCool.ParseDateAfterModification(),
                 TierToArchive = this.TierToArchive is null ? null : this.TierToArchive.ParseDateAfterModification(),
-                Delete = this.Delete is null ? null : this.Delete.ParseDateAfterModification()
+                Delete = this.Delete is null ? null : this.Delete.ParseDateAfterModification(),
+                EnableAutoTierToHotFromCool = this.EnableAutoTierToHotFromCool
             };
         }
     }
@@ -261,20 +264,43 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
     /// </summary>
     public class PSDateAfterModification
     {
-        public int DaysAfterModificationGreaterThan { get; set; }
+        public int? DaysAfterModificationGreaterThan { get; set; }
+        public int? DaysAfterLastAccessTimeGreaterThan { get; set; }
 
-        public PSDateAfterModification(int daysAfterModificationGreaterThan)
+        public PSDateAfterModification()
+        {
+            this.DaysAfterModificationGreaterThan = null;
+            this.DaysAfterLastAccessTimeGreaterThan = null;
+        }
+
+        public PSDateAfterModification(int? daysAfterModificationGreaterThan, int? daysAfterLastAccessTimeGreaterThan)
         {
             this.DaysAfterModificationGreaterThan = daysAfterModificationGreaterThan;
+            this.DaysAfterLastAccessTimeGreaterThan = daysAfterLastAccessTimeGreaterThan;
         }
 
         public PSDateAfterModification(DateAfterModification data)
         {
-            this.DaysAfterModificationGreaterThan = Convert.ToInt32(data.DaysAfterModificationGreaterThan);
+            if (data.DaysAfterModificationGreaterThan is null)
+            {
+                this.DaysAfterModificationGreaterThan = null;
+            }
+            else
+            {
+                this.DaysAfterModificationGreaterThan = Convert.ToInt32(data.DaysAfterModificationGreaterThan);
+            }
+            if (data.DaysAfterLastAccessTimeGreaterThan is null)
+            {
+                this.DaysAfterLastAccessTimeGreaterThan = null;
+            }
+            else
+            {
+                this.DaysAfterLastAccessTimeGreaterThan = Convert.ToInt32(data.DaysAfterLastAccessTimeGreaterThan);
+            }
         }
         public DateAfterModification ParseDateAfterModification()
         {
-            return new DateAfterModification(this.DaysAfterModificationGreaterThan);
+            return new DateAfterModification(this.DaysAfterModificationGreaterThan, this.DaysAfterLastAccessTimeGreaterThan);
         }
     }
 
@@ -284,6 +310,11 @@ namespace Microsoft.Azure.Commands.Management.Storage.Models
     public class PSDateAfterCreation
     {
         public int DaysAfterCreationGreaterThan { get; set; }
+
+        public PSDateAfterCreation()
+        {
+            this.DaysAfterCreationGreaterThan = 0;
+        }
 
         public PSDateAfterCreation(int daysAfterCreationGreaterThan)
         {